# lib/api-zod/src/generated/types/boardInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Quiz Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface BoardInput {
  /** @minLength 1 */
  name: string;
}