# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Quiz Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { QuizSession } from './quizSession';

export interface DashboardStats {
  totalBoards: number;
  totalStandards: number;
  totalSubjects: number;
  totalChapters: number;
  totalTopics: number;
  totalQuizzes: number;
  totalQuestions: number;
  recentQuizzes: QuizSession[];
}