# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Quiz Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionType } from './questionType';

export interface Question {
  id: number;
  question: string;
  type: QuestionType;
  /** @nullable */
  options?: string[] | null;
  correctAnswer: string;
  explanation: string;
  hasDiagram: boolean;
  /** @nullable */
  diagramDescription?: string | null;
  hasEquation: boolean;
  /** @nullable */
  equations?: string[] | null;
  difficulty: string;
}